[NEW_DARTH]Upload asset locally, Able to retranscode asset
Summary:
Added media folder to static path, new upload "post" route, deleting the temp file only when mediaverse asset

Uploading the asset into the existing asset model

Adding an asset lease as owner to the user uploaded the asset

Asset index page contains the assets which are owned by the current user

Re-transcoding the asset

Related to: T2835

Test Plan:
Run "make init" then "make refresh-db" and then "make run" in the terminal at the project directory.
Browse "http://localhost:45000/"

Click on Login with MediaVerse and provide the mediaverse credentials.

Click on "assets" and then choose file to upload and click on 'upload' button.

Once the asset status is 'ready' it can be seen in the asset index page.

Click on "assets" and then click on "Add from MediaVerse".

Click on "add" button beside the mediaverse assets to process and add the assets for fader use.

Once the asset status is 'ready' it can be seen in the asset index page. The button is changed from 'add' to 'added'

Reviewers: stephan.gensch

Reviewed By: stephan.gensch

Subscribers: girish.vadlamudi, ronny.esterluss

Tags: #mediaverse

Differential Revision: https://hive.vrgmnts.net/D1499