[NEW_DARTH] Using Hls to play the dash videos and audio on Chrome and Firefox
Summary:
Displaying the Dimensions as Width and Height

Included the Tailwind UI links

Using svg's as default icons

Using Routes.asset_path instead of url

Using Hls to play the dash videos and audio on Chrome and Firefox

Test Plan:
Run "make init" then "make run" in the terminal at the project directory.
Browse "http://localhost:45000/"

Click on Login with MediaVerse and provide the mediaverse credentials.

Click on assets to and 'add from mediaverse' or 'upload'.

Click on the asset to see the asset details and play the video and audio files on chrome or Firefox

Reviewers: ronny.esterluss

Reviewed By: ronny.esterluss

Subscribers: girish.vadlamudi, stephan.gensch

Tags: #mediaverse

Differential Revision: https://hive.vrgmnts.net/D1507