[New-Darth]Checking asset access for current user
Summary: Checking if the current user has access to the asset to show the asset detail page

Test Plan:
Run "make init" then "make run" in the terminal at the project directory.
Browse "http://localhost:45000/"

Click on Login with MediaVerse and provide the mediaverse credentials.

Click on assets to and 'add from mediaverse' or 'upload'.

After the asset is added, go to asset detail page and copy the url.

Open the application with different browser and login with different account.

Paste the link to see that access is denied.

Reviewers: ronny.esterluss, stephan.gensch

Reviewed By: ronny.esterluss

Subscribers: girish.vadlamudi

Tags: #mediaverse

Differential Revision: https://hive.vrgmnts.net/D1525