[NEW-DARTH] Fader Project Management with project index and detail live_view pages
Summary:
creating project and indexing projects

live_view for project index page

project detail page with list of assets

Project detail page with adding asset to the project option and live_view for this option

Delete project functionality with live_view

functionality to remove project from asset_lease and its live_view

project detail page updates and project thumbnail

Live_view for primary asset lease update in project and live_view for list of assets in project detail page

Using push_navigate when moving from one Live_view to another.

Avoiding the database operation while assigning and unassigning projects to asset_leases

Test Plan:
Run "make init" then "make run" in the terminal at the project directory.
Browse "http://localhost:45000/"

Click on Login with MediaVerse and provide the mediaverse credentials.

Click on assets to and 'add from mediaverse' or 'upload'.

Click on projects to create or see the index of existing projects.

Once the project is created, click on the project to see the project details and the available assets to add to the project.

After the asset is added to the project, there is an option to make it a primary asset to assign that asset's thumbnail image as project thumbnail.

Project can be deleted from the project index page.

Duplicate the browser tab and try to make any changes in the first tab to see the changes in other tab in all these pages.

Reviewers: ronny.esterluss, stephan.gensch

Reviewed By: ronny.esterluss

Subscribers: girish.vadlamudi

Tags: #mediaverse

Differential Revision: https://hive.vrgmnts.net/D1518