[NEW_DARTH] Option to preview mv_assets in mediaverse node
Summary:
Option to preview mv asset in the mv node before adding it to Fader

Restricting asset deletion with changeset

Live_view in mv_asset_index page

Test Plan:
Run "make init", "make refresh-db"(changed some migrations) then "make run" in the terminal at the project directory.
Browse "http://localhost:45000/"

Click on Login with MediaVerse and provide the mediaverse credentials.

Click on assets to and 'add from mediaverse' or 'upload'.

Click on projects to create or see the index of existing projects.

Once the project is created, click on the project to see the project details and the available assets to add to the project.

Try to delete the asset that is added to the project from asset index page. It should fail with the appropriate reason.

Duplicate the browser tab and try to make any changes in the first tab to see the changes in other tab in all these pages.

Reviewers: ronny.esterluss, stephan.gensch

Reviewed By: ronny.esterluss

Subscribers: girish.vadlamudi

Tags: #mediaverse

Differential Revision: https://hive.vrgmnts.net/D1544