[NEW-DARTH] Option to assign asset to the project from asset detail page
Summary:
Project index is available in asset detail page to be able to assign that asset to available projects

DAM provides the username of the user who created the asset under the attribute "createdBy". Displaying it as an extra detail

Test Plan:
Run "make init" then "make run" in the terminal at the project directory.
Browse "http://localhost:45000/"

Click on Login with MediaVerse and provide the mediaverse credentials.

Click on assets to and 'add from mediaverse' or 'upload'.

Click on projects to create or see the index of existing projects.

Once the project is created, click on the assets and select any asset to see the available projects under that asset and an option to add it from there.

Now the asset details in the mv_asset index page contains an extra information about the creator of the asset. Current showing some id but will have the username when there is an update from DAM.

Reviewers: ronny.esterluss, stephan.gensch

Reviewed By: ronny.esterluss

Subscribers: girish.vadlamudi

Tags: #mediaverse

Differential Revision: https://hive.vrgmnts.net/D1547