[NEW_DARTH] Preview of mv_asset
Summary:
Genserver to download the preview assets when user clicks on the mv_asset_index page

Downloading the preview files into folders with preview key name to avoid skipping the download of files with same name

Using image tag for videos as they are in gif codec for preview and not using hls for any media

Broadcasting a message when asset preview is downloaded to refresh the preview page with live_view

Test Plan:
Run "make init" then "make run" in the terminal at the project directory.
Browse "http://localhost:45000/"

Click on Login with MediaVerse and provide the mediaverse credentials.

Click on assets to and click on 'add from mediaverse'

Mediaverse Asset previews are downloaded and visible in the mv_index page with an option to add them to the fader.

Reviewers: ronny.esterluss, stephan.gensch

Reviewed By: ronny.esterluss

Subscribers: girish.vadlamudi

Tags: #mediaverse

Differential Revision: https://hive.vrgmnts.net/D1576