[NEW-DARTH] Fetch MediaVerse projects into Fader
Summary:
When the user decides to add MediaVerse project to Fader. A new project is created in Fader with the project details of the
MediaVerse project and all the assets that are assigned to this Mediaverse project are downloaded to Fader and assigned to the
newly created Fader project.

If the asset status is "created" or "download_failed" there will be a download button instead of re-transcode button on the
asset index card to restart the incomplete download.

Test Plan:
-> Run the application with proxy
-> Login in Mediaverse account
-> Add Mediaverse project to fader

Reviewers: ronny.esterluss, stephan.gensch

Reviewed By: ronny.esterluss

Subscribers: girish.vadlamudi

Tags: #mediaverse

Differential Revision: https://hive.vrgmnts.net/D1654