[NEW-DARTH] Mediaverse asset index page and fader asset index page bug fixes
Summary:
Button "Add all Assets from this page to Fader" previously was just checking the existence of asset based on mv_asset_key. Now along with that check it will create the asset lease for the current user if it does not exist already.

Avoiding the query based on the owner of the asset lease as we are not using it like that anymore

Test Plan:
-> Run application with proxy
-> Login into Fader and add 12 assets to Fader and see that asset index page shows all 12 assets in a single page

Reviewers: ronny.esterluss, stephan.gensch

Reviewed By: ronny.esterluss

Subscribers: girish.vadlamudi

Tags: #mv-wp5_user_driven_immersive_and_accessible_media_authoring

Differential Revision: https://hive.vrgmnts.net/D1701